pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-automation-suite"
version = "0.1.0"
description = "Suite unificada de automação e MCP para o ecossistema Notion do Lucas Biason"
authors = [
    {name = "Lucas Biason", email = "lucas.biason@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["notion", "automation", "mcp", "ai", "workflows"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "pytz>=2024.1",
    "tenacity>=9.0.0",
    "typer>=0.12.5",
    "rich>=13.7.0",
    "requests>=2.32.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0"
]

[project.scripts]
notion-mcp-server = "notion_mcp.server:main"
notion-suite = "notion_automation_suite.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=notion_mcp --cov-report=term-missing"
asyncio_mode = "auto"
pythonpath = ["src"]
